split reservation address refactoring
